+++ featuretools/tests/primitive_tests/test_primitive_base.py
@@ -5,7 +5,7 @@ from ..testing_utils import make_ecommerce_entityset
 
 from featuretools.primitives import Last, Mode, Sum
 from featuretools.primitives.base import Feature, IdentityFeature
-from featuretools.variable_types import Datetime, Numeric
+from featuretools.variable_types import Categorical, Datetime, Id, Numeric
 
 
 @pytest.fixture(scope='module')
@@ -107,3 +107,17 @@ def test_return_type_inference_datetime_time_index(es):
 def test_return_type_inference_numeric_time_index(es_numeric):
     last = Last(es_numeric["log"]["datetime"], es_numeric["customers"])
     assert last.variable_type == Numeric
+
+
+def test_return_type_inference_id(es):
+    # direct features should keep Id variable type
+    direct_id_feature = Feature(es["sessions"]["customer_id"], es["log"])
+    assert direct_id_feature.variable_type == Id
+
+    # aggregations of Id variable types should get converted
+    mode = Mode(es["log"]["session_id"], es["customers"])
+    assert mode.variable_type == Categorical
+
+    # also test direct feature of aggregation
+    mode_direct = Feature(mode, es["sessions"])
+    assert mode_direct.variable_type == Categorical